+++ src/tests/utils.ts
@@ -659,6 +659,7 @@ export function mockEthereumValidator(
       index: index,
       public_key: public_key,
       withdrawal_address: "0xwithdrawal_address_1",
+      withdrawal_credentials: "0x01withdrawal_credentials_1",
       fee_recipient_address: "0xfee_recipient_address_1",
       forwarded_fee_recipient_address: "0xforwarded_fee_recipient_address_1",
       slashed: false,